Remove analysis from compilation
Intel Compiler on Windows 32 failed to compile.
Not wrong code but it seems lot for template and hxx inclusion.
Analysis will be replugged after investigation.

+++ scilab/modules/ast/includes/types/singlestruct.hxx
@@ -50,7 +50,7 @@ public :
 
     inline int getNumFields() const
     {
-        return m_wstFields.size();
+        return static_cast<int>(m_wstFields.size());
     }
 
     bool                                    toString(std::wostringstream& ostr);
